fix: remove integration test of get signed channel
